Update a load of plugins
Doing a `git submodule foreach git pull` will fail if any submodule is
set to a tag or specific commit. Instead, I'm moving to master on each
module even if this means potentially being between releases of a
plugin.

For convenience, here is a list of the changes captured by this parent
commit:
   | plugin             |  commit      | nearest tag |
   |____________________+______________+_____________|
   | ack                |  eede0425f5  |   1.0.9     |
   | dockerfile         |  5d9ad157fa  |   n/a       |
   | vim-characterize   |  c6d26e5017  |   v1.0      |
   | vim-fugitive       |  c00ebd75ac  |   v2.2      |
   | vim-gitgutter      |  b5f23e9867  |   n/a       |
   | vim-go             |  ab01e7e5a6  |   v1.7.1    |
   | vim-json           |  f5e3181d0b  |   n/a       |
   | vim-sort-motion    |  20a599cdda  |   n/a       |
   | vim-surround       |  e49d6c2459  |   v2.1      |
   | vim-troll-stopper  |  b00a4954a2  |   n/a       |
   | vimwiki            |  4831384ab9  |   v2.3      |